# internal/port/http/v1/openapi_type.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package v1

// CreateUserSubRequest defines model for CreateUserSubRequest.
type CreateUserSubRequest struct {
	EndDate     *string `json:"end_date,omitempty"`
	Price       int     `json:"price"`
	ServiceName string  `json:"service_name"`
	StartDate   string  `json:"start_date"`
	UserId      string  `json:"user_id"`
}

// Error defines model for Error.
type Error struct {
	Details string `json:"details"`
	Slug    string `json:"slug"`
}

// UpdateUserSubRequest defines model for UpdateUserSubRequest.
type UpdateUserSubRequest struct {
	EndDate     *string `json:"end_date,omitempty"`
	Price       int     `json:"price"`
	ServiceName string  `json:"service_name"`
	StartDate   string  `json:"start_date"`
	UserId      string  `json:"user_id"`
}

// UserSubscriptionResponse defines model for UserSubscriptionResponse.
type UserSubscriptionResponse struct {
	EndDate     *string `json:"end_date,omitempty"`
	Id          int     `json:"id"`
	Price       int     `json:"price"`
	ServiceName string  `json:"service_name"`
	StartDate   string  `json:"start_date"`
	UserId      string  `json:"user_id"`
}

// UserSubscriptionSumPriceResponse defines model for UserSubscriptionSumPriceResponse.
type UserSubscriptionSumPriceResponse = int

// GetSubsFindByUserParams defines parameters for GetSubsFindByUser.
type GetSubsFindByUserParams struct {
	// UserId user uuid
	UserId string `form:"userId" json:"userId"`
}

// GetSubsGetSumPriceParams defines parameters for GetSubsGetSumPrice.
type GetSubsGetSumPriceParams struct {
	// UserId user uuid
	UserId string `form:"userId" json:"userId"`

	// Service serivce name
	Service string `form:"service" json:"service"`

	// StartDate start period
	StartDate string `form:"startDate" json:"startDate"`

	// EndDate end period
	EndDate string `form:"endDate" json:"endDate"`
}

// CreateUserSubJSONRequestBody defines body for CreateUserSub for application/json ContentType.
type CreateUserSubJSONRequestBody = CreateUserSubRequest

// UpdateSubJSONRequestBody defines body for UpdateSub for application/json ContentType.
type UpdateSubJSONRequestBody = UpdateUserSubRequest